move dev dim reduction and signature detector functions
