+++ hsi_toolkit/dev/qmf_detector.py
