+++ hsi_toolkit/dev/spsmf_detector.py
