+++ hsi_toolkit/dim_reduction/__init__.py
@@ -1,2 +1,1 @@
 from hsi_toolkit.dim_reduction.hdr import *
-from hsi_toolkit.dim_reduction.mnf import *
